If using submit, need new page to show results
Basic PHP forms. When you submit a form to a server, your calculations run on whatever file is set in the "action" attribute, and the results can be printed on that new page. To have something show up on the same page after doing calculations, you need Ajax/JSON, which you don't want to deal with until you get PHP down.